fix: correct concatenation of action sequences
The commit corrects two errors in the concatenation of past and future
actions given a certain policy before doing state estimation. The first
error was the wrong concatenation order of the arrays (past actions were
concatenated after the future ones); the second error was that the
sequence of past actions was sliced until the present time step included,
which is wrong because the action for the current time step has not been
chosen yet (it belongs to the future action sequence).